requirements.txt
cached-property==1.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
dominate==2.7.0
email-validator==1.2.1
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
python-dotenv==0.20.0
requests==2.28.1
SQLAlchemy==1.4.40
sqlalchemy-orm==1.2.3
timeago==1.0.16
typing-inspect==0.7.1
typing_extensions==4.3.0
urllib3==1.26.11
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
